pyproject.toml
[tool.poetry]
name = "votingapp"
version = "0.1.0"
description = ""
authors = ["yanki <yankisirtas@hotmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.5.7"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
SQLAlchemy = "^1.4.2"
psycopg2 = "^2.8.6"
bcrypt = "^3.2.0"
FastAPI-SQLAlchemy = "^0.2.1"
python-dotenv = "^0.15.0"
asyncpg = "^0.22.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.9.0"
bandit = "^1.7.0"
mypy = "^0.812"
isort = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
